+++ exonum/src/lib.rs
@@ -30,9 +30,6 @@ extern crate exonum_profiler;
 extern crate log;
 extern crate byteorder;
 extern crate exonum_sodiumoxide as sodiumoxide;
-#[cfg(feature = "leveldb")]
-extern crate exonum_leveldb as leveldb;
-#[cfg(feature = "rocksdb")]
 extern crate exonum_rocksdb as rocksdb;
 
 extern crate rand;