+++ exonum/src/storage/base_index.rs
@@ -30,7 +30,8 @@ use super::{StorageKey, StorageValue, Snapshot, Fork, Iter};
 /// [`StorageValue`]: ../trait.StorageValue.html
 #[derive(Debug)]
 pub struct BaseIndex<T> {
-    prefix: Vec<u8>,
+    name: String,
+    prefix: Option<Vec<u8>>,
     view: T,
 }
 
@@ -52,25 +53,59 @@ pub struct BaseIndexIter<'a, K, V> {
 }
 
 impl<T> BaseIndex<T> {
-    /// Creates a new index representation based on the common prefix of its keys and storage view.
+    /// Creates a new index representation based on the name and storage view.
     ///
     /// Storage view can be specified as [`&Snapshot`] or [`&mut Fork`]. In the first case only
     /// immutable methods are available. In the second case both immutable and mutable methods are
     /// available.
     /// [`&Snapshot`]: ../trait.Snapshot.html
     /// [`&mut Fork`]: ../struct.Fork.html
-    pub fn new(prefix: Vec<u8>, view: T) -> Self {
-        BaseIndex {
-            prefix: prefix,
-            view: view,
+    pub fn new<S: AsRef<str>>(name: S, view: T) -> Self {
+        if is_valid_name(&name) {
+            BaseIndex {
+                name: name.as_ref().to_string(),
+                prefix: None,
+                view,
+            }
+        } else {
+            panic!("Wrong characters using in name. Use: a-zA-Z0-9 and _");
+        }
+    }
+
+    /// Creates a new index representation based on the name, common prefix of its keys
+    /// and storage view.
+    ///
+    /// Storage view can be specified as [`&Snapshot`] or [`&mut Fork`]. In the first case only
+    /// immutable methods are available. In the second case both immutable and mutable methods are
+    /// available.
+    /// [`&Snapshot`]: ../trait.Snapshot.html
+    /// [`&mut Fork`]: ../struct.Fork.html
+    pub fn with_prefix<S: AsRef<str>>(name: S, prefix: Vec<u8>, view: T) -> Self {
+        if is_valid_name(&name) {
+            BaseIndex {
+                name: name.as_ref().to_string(),
+                prefix: Some(prefix),
+                view,
+            }
+        } else {
+            panic!("Wrong characters using in name. Use: a-zA-Z0-9 and _");
         }
     }
 
     fn prefixed_key<K: StorageKey>(&self, key: &K) -> Vec<u8> {
-        let mut v = vec![0; self.prefix.len() + key.size()];
-        v[..self.prefix.len()].copy_from_slice(&self.prefix);
-        key.write(&mut v[self.prefix.len()..]);
-        v
+        match self.prefix {
+            Some(ref prefix) => {
+                let mut v = vec![0; prefix.len() + key.size()];
+                v[..prefix.len()].copy_from_slice(prefix);
+                key.write(&mut v[prefix.len()..]);
+                v
+            }
+            None => {
+                let mut v = vec![0; key.size()];
+                key.write(&mut v);
+                v
+            }
+        }
     }
 }
 
@@ -84,9 +119,10 @@ where
         K: StorageKey,
         V: StorageValue,
     {
-        self.view.as_ref().get(&self.prefixed_key(key)).map(|v| {
-            StorageValue::from_bytes(Cow::Owned(v))
-        })
+        self.view
+            .as_ref()
+            .get(&self.name, &self.prefixed_key(key))
+            .map(|v| StorageValue::from_bytes(Cow::Owned(v)))
     }
 
     /// Returns `true` if the index contains a value of *any* type for the specified key of
@@ -95,7 +131,10 @@ where
     where
         K: StorageKey,
     {
-        self.view.as_ref().contains(&self.prefixed_key(key))
+        self.view.as_ref().contains(
+            &self.name,
+            &self.prefixed_key(key),
+        )
     }
 
     /// Returns an iterator over the entries of the index in ascending order. The iterator element
@@ -109,8 +148,8 @@ where
     {
         let iter_prefix = self.prefixed_key(subprefix);
         BaseIndexIter {
-            base_iter: self.view.as_ref().iter(&iter_prefix),
-            base_prefix_len: self.prefix.len(),
+            base_iter: self.view.as_ref().iter(&self.name, &iter_prefix),
+            base_prefix_len: self.prefix.as_ref().map_or(0, |p| p.len()),
             prefix: iter_prefix,
             ended: false,
             _k: PhantomData,
@@ -131,8 +170,8 @@ where
         let iter_prefix = self.prefixed_key(subprefix);
         let iter_from = self.prefixed_key(from);
         BaseIndexIter {
-            base_iter: self.view.as_ref().iter(&iter_from),
-            base_prefix_len: self.prefix.len(),
+            base_iter: self.view.as_ref().iter(&self.name, &iter_from),
+            base_prefix_len: self.prefix.as_ref().map_or(0, |p| p.len()),
             prefix: iter_prefix,
             ended: false,
             _k: PhantomData,
@@ -149,7 +188,7 @@ impl<'a> BaseIndex<&'a mut Fork> {
         V: StorageValue,
     {
         let key = self.prefixed_key(key);
-        self.view.put(key, value.into_bytes());
+        self.view.put(&self.name, key, value.into_bytes());
     }
 
     /// Removes the key of *any* type from the index.
@@ -158,10 +197,11 @@ impl<'a> BaseIndex<&'a mut Fork> {
         K: StorageKey,
     {
         let key = self.prefixed_key(key);
-        self.view.remove(key);
+        self.view.remove(&self.name, key);
     }
 
-    /// Clears the index, removing all entries.
+    /// Clears the index, removing entries with keys that starts with a prefix or all entries
+    /// if `prefix` is `None`.
     ///
     /// # Notes
     ///
@@ -169,7 +209,7 @@ impl<'a> BaseIndex<&'a mut Fork> {
     /// this method the amount of allocated memory is linearly dependent on the number of elements
     /// in the index.
     pub fn clear(&mut self) {
-        self.view.remove_by_prefix(&self.prefix)
+        self.view.remove_by_prefix(&self.name, self.prefix.as_ref());
     }
 }
 
@@ -201,4 +241,30 @@ impl<'a, K, V> ::std::fmt::Debug for BaseIndexIter<'a, K, V> {
     fn fmt(&self, f: &mut ::std::fmt::Formatter) -> ::std::fmt::Result {
         write!(f, "BaseIndexIter(..)")
     }
+}
+
+/// A function that validates an index name. Allowable characters in name: ASCII characters, digits
+/// and underscores.
+fn is_valid_name<S: AsRef<str>>(name: S) -> bool {
+    name.as_ref().as_bytes().iter().all(|c| match *c {
+        48...57 | 65...90 | 97...122 | 95 | 46 => true,
+        _ => false,
+    })
+}
+
+#[test]
+fn test_index_name_validator() {
+    assert!(is_valid_name("index_name"));
+    assert!(is_valid_name("_index_name"));
+    assert!(is_valid_name("AinDex_name_"));
+    assert!(is_valid_name("core.index_name1Z"));
+    assert!(is_valid_name("configuration.indeX_1namE"));
+    assert!(is_valid_name("1index_Namez"));
+
+    assert!(!is_valid_name("index-name"));
+    assert!(!is_valid_name("_index-name"));
+    assert!(!is_valid_name("индекс_name_"));
+    assert!(!is_valid_name("core.index_имя3"));
+    assert!(!is_valid_name("indeX_1namE-"));
+    assert!(!is_valid_name("1in!dex_Namez"));
 }