+++ exonum/src/storage/mod.rs
@@ -112,10 +112,7 @@
 pub use self::error::Error;
 pub use self::db::{Database, Snapshot, Fork, Patch, Change, Iterator, Iter};
 
-#[cfg(feature = "leveldb")]
-pub use self::leveldb::{LevelDB, LevelDBOptions, LevelDBCache};
-#[cfg(feature = "rocksdb")]
-pub use self::rocksdb::{RocksDB, RocksDBOptions, RocksBlockOptions};
+pub use self::rocksdb::{RocksDB, RocksDBOptions};
 pub use self::memorydb::MemoryDB;
 
 pub use self::keys::StorageKey;
@@ -137,9 +134,6 @@ pub type Result<T> = ::std::result::Result<T, Error>;
 
 mod error;
 mod db;
-#[cfg(feature = "leveldb")]
-mod leveldb;
-#[cfg(feature = "rocksdb")]
 mod rocksdb;
 mod memorydb;