+++ exonum/src/storage/proof_list_index/tests.rs
@@ -22,6 +22,8 @@ use encoding::serialize::reexport::Serialize;
 
 use self::ListProof::*;
 
+const IDX_NAME: &'static str = "idx_name";
+
 fn random_values(len: usize) -> Vec<Vec<u8>> {
     use std::collections::HashSet;
     let mut rng = thread_rng();
@@ -46,7 +48,7 @@ fn gen_tempdir_name() -> String {
 
 fn list_methods(db: Box<Database>) {
     let mut fork = db.fork();
-    let mut index = ProofListIndex::new(vec![255], &mut fork);
+    let mut index = ProofListIndex::new(IDX_NAME, &mut fork);
 
     assert!(index.is_empty());
     assert_eq!(index.len(), 0);
@@ -67,7 +69,7 @@ fn list_methods(db: Box<Database>) {
 
 fn height(db: Box<Database>) {
     let mut fork = db.fork();
-    let mut index = ProofListIndex::new(vec![255], &mut fork);
+    let mut index = ProofListIndex::new(IDX_NAME, &mut fork);
 
     index.push(vec![1]);
     assert_eq!(index.height(), 1);
@@ -93,7 +95,7 @@ fn height(db: Box<Database>) {
 
 fn iter(db: Box<Database>) {
     let mut fork = db.fork();
-    let mut list_index = ProofListIndex::new(vec![255], &mut fork);
+    let mut list_index = ProofListIndex::new(IDX_NAME, &mut fork);
 
     list_index.extend(vec![1u8, 2, 3]);
 
@@ -108,7 +110,7 @@ fn iter(db: Box<Database>) {
 
 fn list_index_proof(db: Box<Database>) {
     let mut fork = db.fork();
-    let mut index = ProofListIndex::new(vec![255], &mut fork);
+    let mut index = ProofListIndex::new(IDX_NAME, &mut fork);
 
     let h0 = 2u64.hash();
     let h1 = 4u64.hash();
@@ -247,7 +249,7 @@ fn list_index_proof(db: Box<Database>) {
 
 fn randomly_generate_proofs(db: Box<Database>) {
     let mut fork = db.fork();
-    let mut index = ProofListIndex::new(vec![255], &mut fork);
+    let mut index = ProofListIndex::new(IDX_NAME, &mut fork);
     let num_vals = 100;
     let values = random_values(num_vals as usize);
     let mut rng = thread_rng();
@@ -291,7 +293,7 @@ fn randomly_generate_proofs(db: Box<Database>) {
 
 fn index_and_proof_roots(db: Box<Database>) {
     let mut fork = db.fork();
-    let mut index = ProofListIndex::new(vec![255], &mut fork);
+    let mut index = ProofListIndex::new(IDX_NAME, &mut fork);
     assert_eq!(index.root_hash(), Hash::zero());
 
     let h1 = hash(&[1, 2]);
@@ -419,14 +421,14 @@ fn index_and_proof_roots(db: Box<Database>) {
 
 fn proof_illegal_lower_bound(db: Box<Database>) {
     let mut fork = db.fork();
-    let mut index = ProofListIndex::new(vec![255], &mut fork);
+    let mut index = ProofListIndex::new(IDX_NAME, &mut fork);
     index.get_range_proof(0, 1);
     index.push(vec![1]);
 }
 
 fn proof_illegal_bound_empty(db: Box<Database>) {
     let mut fork = db.fork();
-    let mut index = ProofListIndex::new(vec![255], &mut fork);
+    let mut index = ProofListIndex::new(IDX_NAME, &mut fork);
     for i in 0u8..8 {
         index.push(vec![i]);
     }
@@ -435,7 +437,7 @@ fn proof_illegal_bound_empty(db: Box<Database>) {
 
 fn proof_illegal_range(db: Box<Database>) {
     let mut fork = db.fork();
-    let mut index = ProofListIndex::new(vec![255], &mut fork);
+    let mut index = ProofListIndex::new(IDX_NAME, &mut fork);
     for i in 0u8..4 {
         index.push(vec![i]);
     }
@@ -444,7 +446,7 @@ fn proof_illegal_range(db: Box<Database>) {
 
 fn proof_structure(db: Box<Database>) {
     let mut fork = db.fork();
-    let mut index = ProofListIndex::new(vec![255], &mut fork);
+    let mut index = ProofListIndex::new(IDX_NAME, &mut fork);
     assert_eq!(index.root_hash(), Hash::zero());
 
     let h1 = hash(&vec![0, 1, 2]);
@@ -496,7 +498,7 @@ fn simple_root_hash(db: Box<Database>) {
     let h2 = hash(&[2]);
 
     let mut fork = db.fork();
-    let mut index = ProofListIndex::new(vec![255], &mut fork);
+    let mut index = ProofListIndex::new(IDX_NAME, &mut fork);
     assert_eq!(index.get(0), None);
     index.push(vec![1]);
     assert_eq!(index.root_hash(), h1);
@@ -508,7 +510,7 @@ fn simple_root_hash(db: Box<Database>) {
 fn same_root_hash(db1: Box<Database>, db2: Box<Database>) {
     let mut fork1 = db1.fork();
 
-    let mut i1 = ProofListIndex::new(vec![255], &mut fork1);
+    let mut i1 = ProofListIndex::new(IDX_NAME, &mut fork1);
     i1.push(vec![1]);
     i1.push(vec![2]);
     i1.push(vec![3]);
@@ -521,7 +523,7 @@ fn same_root_hash(db1: Box<Database>, db2: Box<Database>) {
 
     let mut fork2 = db2.fork();
 
-    let mut i2 = ProofListIndex::new(vec![255], &mut fork2);
+    let mut i2 = ProofListIndex::new(IDX_NAME, &mut fork2);
     i2.push(vec![4]);
     i2.push(vec![7]);
     i2.push(vec![5]);
@@ -651,122 +653,6 @@ mod memorydb_tests {
     }
 }
 
-#[cfg(feature = "leveldb")]
-mod leveldb_tests {
-    use std::path::Path;
-    use tempdir::TempDir;
-    use storage::{Database, LevelDB, LevelDBOptions};
-
-    fn create_database(path: &Path) -> Box<Database> {
-        let mut opts = LevelDBOptions::default();
-        opts.create_if_missing = true;
-        Box::new(LevelDB::open(path, opts).unwrap())
-    }
-
-    #[test]
-    fn test_list_methods() {
-        let dir = TempDir::new(super::gen_tempdir_name().as_str()).unwrap();
-        let path = dir.path();
-        let db = create_database(path);
-        super::list_methods(db);
-    }
-
-    #[test]
-    fn test_height() {
-        let dir = TempDir::new(super::gen_tempdir_name().as_str()).unwrap();
-        let path = dir.path();
-        let db = create_database(path);
-        super::height(db);
-    }
-
-    #[test]
-    fn test_iter() {
-        let dir = TempDir::new(super::gen_tempdir_name().as_str()).unwrap();
-        let path = dir.path();
-        let db = create_database(path);
-        super::iter(db);
-    }
-
-    #[test]
-    fn test_list_index_proof() {
-        let dir = TempDir::new(super::gen_tempdir_name().as_str()).unwrap();
-        let path = dir.path();
-        let db = create_database(path);
-        super::list_index_proof(db);
-    }
-
-    #[test]
-    fn test_randomly_generate_proofs() {
-        let dir = TempDir::new(super::gen_tempdir_name().as_str()).unwrap();
-        let path = dir.path();
-        let db = create_database(path);
-        super::randomly_generate_proofs(db);
-    }
-
-    #[test]
-    fn test_index_and_proof_roots() {
-        let dir = TempDir::new(super::gen_tempdir_name().as_str()).unwrap();
-        let path = dir.path();
-        let db = create_database(path);
-        super::index_and_proof_roots(db);
-    }
-
-    #[test]
-    #[should_panic]
-    fn test_proof_illegal_lower_bound() {
-        let dir = TempDir::new(super::gen_tempdir_name().as_str()).unwrap();
-        let path = dir.path();
-        let db = create_database(path);
-        super::proof_illegal_lower_bound(db);
-    }
-
-    #[test]
-    #[should_panic]
-    fn test_proof_illegal_bound_empty() {
-        let dir = TempDir::new(super::gen_tempdir_name().as_str()).unwrap();
-        let path = dir.path();
-        let db = create_database(path);
-        super::proof_illegal_bound_empty(db);
-    }
-
-    #[test]
-    #[should_panic]
-    fn test_proof_illegal_range() {
-        let dir = TempDir::new(super::gen_tempdir_name().as_str()).unwrap();
-        let path = dir.path();
-        let db = create_database(path);
-        super::proof_illegal_range(db);
-    }
-
-    #[test]
-    fn test_proof_structure() {
-        let dir = TempDir::new(super::gen_tempdir_name().as_str()).unwrap();
-        let path = dir.path();
-        let db = create_database(path);
-        super::proof_structure(db);
-    }
-
-    #[test]
-    fn test_siple_root_hash() {
-        let dir = TempDir::new(super::gen_tempdir_name().as_str()).unwrap();
-        let path = dir.path();
-        let db = create_database(path);
-        super::simple_root_hash(db);
-    }
-
-    #[test]
-    fn test_same_root_hash() {
-        let dir1 = TempDir::new(super::gen_tempdir_name().as_str()).unwrap();
-        let path1 = dir1.path();
-        let db1 = create_database(path1);
-        let dir2 = TempDir::new(super::gen_tempdir_name().as_str()).unwrap();
-        let path2 = dir2.path();
-        let db2 = create_database(path2);
-        super::same_root_hash(db1, db2);
-    }
-}
-
-#[cfg(feature = "rocksdb")]
 mod rocksdb_tests {
     use std::path::Path;
     use tempdir::TempDir;